cmd/contour: add IngressStatusWriter
Updates #403

This PR adds a worker to doServe's group which writes back load balancer
status to ingress.v1beta1 objects.

In this PR, the population of the isw.lbStatus channel is not present
but has been manually verified with the following code:

        g.Add(func(stop <-chan struct{}) error {
                tick := time.Tick(2 * time.Second)
                for {
                        select {
                        case <-tick:
                                isw.lbStatus <- v1.LoadBalancerStatus{
                                        Ingress: []v1.LoadBalancerIngress{{
                                                Hostname: "gke.davecheney.com",
                                        }},
                                }
                        case <-stop:
                                return nil
                        }
                }
        })

Future PRs will add workers to populate the channel.

Signed-off-by: Dave Cheney <dave@cheney.net>